Major refactor into using BigQuerySource
Start to pull the logic together into a more generic job framework.
Still need parameterization and loading of the Graph model and table
names, but most of the plumbing for using those elements is now there.

Had to adopt a similar design pattern to the Beam variant, i.e.
generating tuples of table name, stream name to make sure we could
update the metadata in the Arrow objects.