fix bug when appending to BQ with very small data.
If a single batch of data was being read from GDS, it's possible a
PySpark worker may end up with an empty iterable. This was triggering an
attempt to finalize a stream...but there was never an append attempted
so there is no stream! Was causing a runtime error even though data
*was* being appended by other workers.

Bump the patch version as a result.